refactor(planning_test_utils): separate hpp and cpp (#7082)
